requirements.txt
# Personal Agent Team - Complete Dependencies
# This file contains all dependencies for both agents and the production API

# Core Agent Dependencies
agno>=0.1.0
firecrawl>=0.1.0
anthropic>=0.8.0
openai>=1.0.0
lancedb>=0.6.0
pypdf>=4.0.0
python-dotenv>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
tantivy>=0.20.0
pylance>=0.25.2
yfinance>=0.2.28
python-telegram-bot>=21.0.0

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication and security (optional)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Monitoring and logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Scheduling
apscheduler>=3.10.0

# HTTP client for external APIs
httpx>=0.25.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0